requirements.txt
google-adk==0.1.0
yfinance==0.2.65
